+++ src/flatten.js
@@ -14,21 +14,25 @@ function flattenOnce ( resolved ) {
 	let flattened = []
 
 	resolved.forEach(( item ) => {
-		if (item.type != "Group") {
-			return flattened.push(item)
-		}
+		switch (item.type) {
+			case "Group": {
+				let { extend, ...group } = item
 
-		let { extend, ...group } = item
+				flattened.push({
+					...group,
+					extend: []
+				})
 
-		flattened.push({
-			...group,
-			extend: []
-		})
+				if (extend.length) {
+					let more = flattenExtend(extend, group.selectors)
 
-		if (extend.length) {
-			let more = flattenExtend(extend, group.selectors)
-
-			flattened = flattened.concat(more)
+					flattened = flattened.concat(more)
+				}
+				break
+			}
+			default: {
+				flattened.push(item)
+			}
 		}
 	})
 
@@ -36,6 +40,17 @@ function flattenOnce ( resolved ) {
 }
 
 export default function flatten ( resolved ) {
+	resolved = resolved.map(({ tokens, ...item }) => {
+		if (tokens) {
+			return {
+				...item,
+				tokens: flatten(tokens)
+			}
+		} else {
+			return item
+		}
+	})
+
 	while (resolved.filter(({ extend }) => extend?.length).length) {
 		resolved = flattenOnce(resolved)
 	}