+++ src/resolve.js
@@ -1,14 +1,27 @@
 import { lookup } from "./_lookup.js"
+import { warn } from "./_utils.js"
 
-function resolveProperty ({ type, name, props, lineNumber }, top ) {
+function resolveArgs ( params, args, lineNumber ) {
+	if (args.length > params.length) warn(`line: ${lineNumber}: expected ${params.length} arguments but got ${args.length}`)
+
+	let resolved = {}
+	params.forEach(([ param, def ], i) => {
+		if (!def && !args[i]) throw { message: `expected ${i} or more arguments but got only ${args.length}`, lineNumber }
+
+		resolved[param] = args[i] || def
+	})
+
+	return resolved
+}
+
+function resolveProperty ({ type, name, props, lineNumber }, top, args = {} ) {
 	let nName = lookup(name)
 	props = props.map(( prop ) => {
 		if (/^\$/.test(prop)) {
-			let value = top.find(({ type, name: n }) => type == "VariableAssignment" && n == prop)
-
+			let value = args[prop] || top.find(({ type, name: n }) => type == "VariableAssignment" && n == prop)?.value
 			if (!value) throw { message: `variable ${prop} does not exist`, lineNumber }
 
-			return value.value
+			return value
 		} else return prop
 	})
 
@@ -22,13 +35,14 @@ function resolveProperty ({ type, name, props, lineNumber }, top ) {
 
 function resolveMixin ({ name, args, lineNumber }, top ) {
 	let mix = top.find(({ type, name: n }) => type == "MixinDeclaration" && n == name)
-
 	if (!mix) throw { message: `mixin ${name} does not exist.`, lineNumber }
 
-	return resolveGroup(mix, top).properties
+	let resolved = resolveArgs(mix.parameters, args, lineNumber)
+
+	return resolveGroup(mix, top, resolved).properties
 }
 
-function resolveGroup ( group, top ) {
+function resolveGroup ( group, top, args = {} ) {
 	let { properties, ...rest } = group
 
 	let nProps = []
@@ -44,7 +58,7 @@ function resolveGroup ( group, top ) {
 			}
 			case "Property": {
 				nProps.push(
-					resolveProperty(prop, top)
+					resolveProperty(prop, top, args)
 				)
 
 				break
@@ -64,7 +78,7 @@ function resolveGroup ( group, top ) {
 	}
 }
 
-export default function resolve ( tokens ) {
+export default function resolve ( tokens, top ) {
 	let resolved = []
 
 	let index = 0
@@ -72,13 +86,23 @@ export default function resolve ( tokens ) {
 		let token = tokens[index]
 
 		switch (token.type) {
-			case "Special": {
+			case "RegularAt": {
 				resolved.push(token)
+				break
+			}
+			case "NestedAt": {
+				let { tokens: nTokens, ...rest } = token
+
+				resolved.push({
+					...rest,
+					tokens: resolve(nTokens, top ?? tokens.slice(0, index).reverse())
+				})
+
 				break
 			}
 			case "Group": {
 				resolved.push(
-					resolveGroup(token, tokens.slice(0, index).reverse())
+					resolveGroup(token, top ?? tokens.slice(0, index).reverse())
 				)
 				break
 			}