+++ src/tokenize.js
@@ -1,7 +1,35 @@
-function createSelectors ( lines, index ) {
+function resolveMixinParameters ( unparsed, lineNumber ) {
+	if (!unparsed) return []
+	unparsed = unparsed.replace(/^\(?\s+|\s+\)?$/g, "")
+
+	let params = []
+
+	let parsed = unparsed.split(/\s+/g).map(( el ) => el.split(/\s*(,|=)\s*/)).flat().filter(( el ) => el.length)
+	parsed.forEach(( param, i ) => {
+		let last = parsed[i - 1]
+		if (!last && /[,=]/.test(param)) throw { message: `parameters may not start with ${param}`, lineNumber }
+
+		if (param.startsWith("$")) {
+			if (!/^\$[a-z][a-z0-9]*$/.test(param)) throw { message: "malformed variable in mixin parameters", lineNumber }
+			if (last && last == "=") throw { message: "cannot assign default variable to a variable", lineNumber }
+
+			params.push([ param ])
+		} else if (!/[,=]/.test(param)) {
+			if (!params.length) throw { message: "cannot set a default value if no variable has been declared in mixin parameters", lineNumber }
+			if (!/[,=]/.test(last) && !/^\$[a-z][a-z0-9]*$/.test(last)) throw { message: "default may only be one value", lineNumber }
+			if (last == ",") throw { message: "comma cannot follow before default value declaration", lineNumber }
+
+			params[params.length - 1][1] = param
+		}
+	})
+
+	return params
+}
+
+function createSelectors ( lines, index, level ) {
 	let selectors = []
 
-	while (!lines[index].line.startsWith("\t")) {
+	while (!lines[index]?.line.startsWith("\t".repeat(level + 1))) {
 		let selector = lines[index].line.trim().replace(/,+$/, "").split(/,/).map(( selector ) => selector.trim()).filter(( selector ) => selector)
 		selectors.push(...selector)
 
@@ -16,19 +44,18 @@ function createSelectors ( lines, index ) {
 
 function createProperties ( lines, index, level ) {
 	let property = []
-	let indentation = Array(level).fill("\t").join("")
+	let indentation = "\t".repeat(level)
 
 	while (lines[index] && lines[index].line.startsWith(indentation)) {
 		let { line, lineNumber } = lines[index]
 		line = line.trim()
 
 		if (line.startsWith("%")) {
-			let match = /^\% *(?<name>[a-z][a-z0-9]*) *(\((?<args>.+)\))? *$/i.exec(line)?.groups
-
+			let match = /^\% *([a-z][a-z0-9]*) *(\(.+\)|.+)? *$/i.exec(line)
 			if (!match) throw { message: "malformed mixin", lineNumber }
 
-			let { name, args = "" } = match
-			args = args.trim().split(/[, ]+/g).filter(( arg ) => arg.length)
+			let [, name, args = "" ] = match
+			args = args.replace(/^\(\s*|\s*\)$/g, "").split(/[, ]+/g).filter(( arg ) => arg.length)
 
 			property.push({
 				type: "UseMixin",
@@ -40,7 +67,6 @@ function createProperties ( lines, index, level ) {
 			index++
 		} else if (line.startsWith("&")) {
 			if (!/^& *[^ ]+.*$/.test(line)) throw { message: "malformed extend stuff", lineNumber }
-
 			let [, selectors ] = /^& *(.+)$/.exec(line)
 
 			let { index: pIndex, properties: props } = createProperties(lines, index + 1, level + 1)
@@ -54,25 +80,26 @@ function createProperties ( lines, index, level ) {
 				lineNumber
 			})
 		} else {
-			let match = /^([a-z\-]+) *[: ] *(.+?);?$|^(\$[a-z][a-z0-9]*)$/gi.exec(line)
-
+			let match = /^([a-z\-]+) *[: ] *(.+?);?$|^(\$[a-z][a-z0-9]*)( +.+)?$/gi.exec(line)
 			if (!match) throw { message: "malformed property", lineNumber }
 
-			let [, name, props, single ] = match
-			if (!single) {
-				props = props.trim().split(/ +/g)
-				
+			let [, regular, regularProps, shortcut, shortcutProps ] = match
+			if (!shortcut) {
+				let props = regularProps.trim().split(/ +/g)
+
 				property.push({
 					type: "Property",
-					name,
+					name: regular,
 					props,
 					lineNumber
 				})
 			} else {
+				let props = shortcutProps?.trim() ? shortcutProps.trim().split(/ +/g) : []
+
 				property.push({
 					type: "Property",
-					name: single.slice(1),
-					props: [ single ],
+					name: shortcut.slice(1),
+					props: [ shortcut, ...props ],
 					lineNumber
 				})
 			}
@@ -87,18 +114,17 @@ function createProperties ( lines, index, level ) {
 	}
 }
 
-export default function tokenize ( file ) {
-	let lines = file.split(/\r?\n/g).map(( line ) => line.trimEnd()).map(( line, number ) => ({ line, lineNumber: number + 1 })).filter(({ line }) => line)
+const regularAt = [ "charset", "import", "namespace" ]
 
+function createGroups ( lines, index, level ) {
 	let tokens = []
 
-	let index = 0
-	while (index < lines.length) {
+	while (index < lines.length && lines[index].line.startsWith("\t".repeat(level))) {
 		let { line, lineNumber } = lines[index]
+		line = line.slice(level)
 
 		if (/^\$/.test(line)) {
 			let match = /^(?<name>\$[a-z][a-z0-9]*)( *= *| +)(?<value>.+) *$/i.exec(line)?.groups
-
 			if (!match) throw { message: "malformed variable assignment", lineNumber }
 
 			tokens.push({
@@ -109,36 +135,60 @@ export default function tokenize ( file ) {
 
 			index++
 		} else if (/^%/.test(line)) {
-			let match = /^\% *(?<name>[a-z][a-z0-9]*)/i.exec(line)?.groups
-
+			let match = /^\% *([a-z][a-z0-9]*) *([^()]+|\(.+\))? *$/i.exec(line)
 			if (!match) throw { message: "malformed mixin declaration", lineNumber }
+
+			let [, name, params ] = match
+			let parameters = resolveMixinParameters(params)
 
 			let { index: pIndex, properties } = createProperties(lines, index + 1, 1)
 			index = pIndex
 
 			tokens.push({
 				type: "MixinDeclaration",
-				name: match.name,
+				name,
+				parameters,
 				properties,
 				lineNumber
 			})
 		} else if (/^@/.test(line)) {
-			let match = /^@(?<keyword>[a-z]+) *(?<props>.+)? *$/i.exec(line)?.groups
+			let match = /^@([a-z\-]+)/i.exec(line)
+			if (!match) throw { message: "malformed @-rule", lineNumber }
 
-			if (!match) throw { message: "malformed special stuff", lineNumber }
+			let [, name ] = match
+			if (regularAt.includes(name)) {
+				let rMatch = /^@[a-z\-]+ +(.+);?/i.exec(line)
+				if (!rMatch) throw { message: "malformed @-rule", lineNumber }
 
-			tokens.push({
-				type: "Special",
-				...match,
-				lineNumber
-			})
+				let [, rule ] = rMatch
 
-			index++
+				tokens.push({
+					type: "RegularAt",
+					name,
+					rule: rule.trim(),
+					lineNumber
+				})
+
+				index++
+			} else {
+				let [, rule ] = /^@[^ ]+ +(.+)?/i.exec(line) || []
+
+				let { tokens: nTokens, index: nIndex } = createGroups(lines, index + 1, level + 1)
+				index = nIndex
+
+				tokens.push({
+					type: "NestedAt",
+					name,
+					rule,
+					tokens: nTokens,
+					lineNumber
+				})
+			}
 		} else {
-			let { index: sIndex, selectors } = createSelectors(lines, index)
+			let { index: sIndex, selectors } = createSelectors(lines, index, level)
 			index = sIndex
 
-			let { index: pIndex, properties } = createProperties(lines, index, 1)
+			let { index: pIndex, properties } = createProperties(lines, index, level + 1)
 			index = pIndex
 
 			tokens.push({
@@ -149,6 +199,22 @@ export default function tokenize ( file ) {
 			})
 		}
 	}
+
+	return {
+		tokens,
+		index
+	}
+}
+
+export default function tokenize ( file ) {
+	let lines = file
+		.split(/\r?\n/g)
+		.map(( line ) => line.replace(/\/\/.+$/, ""))
+		.map(( line ) => line.trimEnd())
+		.map(( line, number ) => ({ line, lineNumber: number + 1 }))
+		.filter(({ line }) => line)
+
+	let { tokens } = createGroups(lines, 0, 0)
 
 	return tokens
 }